# web/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Voice as a Service",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/agents/schema": {
   "get": {
    "tags": [
     "agents"
    ],
    "summary": "Definition Schema",
    "description": "JSON Schema de la definicion de un agente (para el editor de la UI).",
    "operationId": "definition_schema_api_v1_agents_schema_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Definition Schema Api V1 Agents Schema Get"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/agents/validate": {
   "post": {
    "tags": [
     "agents"
    ],
    "summary": "Validate Definition",
    "description": "Valida una definicion sin guardarla. id y version los pone la app.",
    "operationId": "validate_definition_api_v1_agents_validate_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/DefinitionIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ValidationOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/agents/prompt": {
   "post": {
    "tags": [
     "agents"
    ],
    "summary": "Preview Prompt",
    "description": "El prompt que arma el motor con esta definicion (sin guardarla): el system prompt sale\nsiempre de la definicion. classic: todo el agente en el prompt de sistema; structured: un\nprompt de sistema fijo y la definicion en cada turno.",
    "operationId": "preview_prompt_api_v1_agents_prompt_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PromptIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PromptOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/agent-templates": {
   "get": {
    "tags": [
     "agents"
    ],
    "summary": "List Templates",
    "operationId": "list_templates_api_v1_agent_templates_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/TemplateOut"
         },
         "type": "array",
         "title": "Response List Templates Api V1 Agent Templates Get"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/agent-templates/{template_id}": {
   "get": {
    "tags": [
     "agents"
    ],
    "summary": "Get Template",
    "operationId": "get_template_api_v1_agent_templates__template_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "template_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Template Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Get Template Api V1 Agent Templates  Template Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/agents": {
   "get": {
    "tags": [
     "agents"
    ],
    "summary": "List Agents",
    "operationId": "list_agents_api_v1_agents_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "client_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Client Id"
      }
     },
     {
      "name": "include_archived",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Include Archived"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/AgentOut"
         },
         "title": "Response List Agents Api V1 Agents Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "agents"
    ],
    "summary": "Create Agent",
    "description": "Admin: en el cliente de client_id. Usuario o API key de un cliente: en el suyo\n(client_id se puede omitir; el de otro cliente da 404).",
    "operationId": "create_agent_api_v1_agents_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AgentCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AgentDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/agents/{agent_id}": {
   "get": {
    "tags": [
     "agents"
    ],
    "summary": "Read Agent",
    "operationId": "read_agent_api_v1_agents__agent_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "agent_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Agent Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AgentDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "patch": {
    "tags": [
     "agents"
    ],
    "summary": "Update Agent",
    "operationId": "update_agent_api_v1_agents__agent_id__patch",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "agent_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Agent Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AgentUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AgentDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "agents"
    ],
    "summary": "Delete Agent",
    "operationId": "delete_agent_api_v1_agents__agent_id__delete",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "agent_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Agent Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/agents/{agent_id}/definition": {
   "put": {
    "tags": [
     "agents"
    ],
    "summary": "Update Definition",
    "description": "Guarda la definicion como version nueva (si cambio). Las llamadas en curso\nsiguen con la version con que empezaron.",
    "operationId": "update_definition_api_v1_agents__agent_id__definition_put",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "agent_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Agent Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/DefinitionIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AgentDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/agents/{agent_id}/versions": {
   "get": {
    "tags": [
     "agents"
    ],
    "summary": "List Versions",
    "operationId": "list_versions_api_v1_agents__agent_id__versions_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "agent_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Agent Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/AgentVersionOut"
         },
         "title": "Response List Versions Api V1 Agents  Agent Id  Versions Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/agents/{agent_id}/versions/{version}": {
   "get": {
    "tags": [
     "agents"
    ],
    "summary": "Read Version",
    "operationId": "read_version_api_v1_agents__agent_id__versions__version__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "agent_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Agent Id"
      }
     },
     {
      "name": "version",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Version"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AgentVersionOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
   "AgentCreate": {
    "properties": {
     "client_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Client Id",
      "description": "Obligatorio para un admin; un usuario o API key de cliente crea en el suyo"
     },
     "name": {
      "type": "string",
      "maxLength": 128,
      "minLength": 1,
      "title": "Name"
     },
     "slug": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^[a-z0-9][a-z0-9_]{0,63}$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Slug",
      "description": "Sin slug, se arma del nombre"
     },
     "description": {
      "type": "string",
      "title": "Description",
      "default": ""
     },
     "definition": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Definition"
     },
     "template_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Template Id"
     },
     "engine": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "classic",
         "structured"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Engine",
      "description": "Motor; pisa el de la definicion o la plantilla (la definicion es la misma)"
     }
    },
    "type": "object",
    "required": [
     "name"
    ],
    "title": "AgentCreate"
   },
   "AgentDetail": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "client_id": {
      "type": "string",
      "title": "Client Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "slug": {
      "type": "string",
      "title": "Slug"
     },
     "description": {
      "type": "string",
      "title": "Description"
     },
     "version": {
      "type": "integer",
      "title": "Version"
     },
     "archived": {
      "type": "boolean",
      "title": "Archived"
     },
     "engine": {
      "type": "string",
      "title": "Engine"
     },
     "voice": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Voice"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "updated_at": {
      "type": "string",
      "format": "date-time",
      "title": "Updated At"
     },
     "definition": {
      "additionalProperties": true,
      "type": "object",
      "title": "Definition"
     }
    },
    "type": "object",
    "required": [
     "id",
     "client_id",
     "name",
     "slug",
     "description",
     "version",
     "archived",
     "engine",
     "voice",
     "created_at",
     "updated_at",
     "definition"
    ],
    "title": "AgentDetail"
   },
   "AgentOut": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "client_id": {
      "type": "string",
      "title": "Client Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "slug": {
      "type": "string",
      "title": "Slug"
     },
     "description": {
      "type": "string",
      "title": "Description"
     },
     "version": {
      "type": "integer",
      "title": "Version"
     },
     "archived": {
      "type": "boolean",
      "title": "Archived"
     },
     "engine": {
      "type": "string",
      "title": "Engine"
     },
     "voice": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Voice"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "updated_at": {
      "type": "string",
      "format": "date-time",
      "title": "Updated At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "client_id",
     "name",
     "slug",
     "description",
     "version",
     "archived",
     "engine",
     "voice",
     "created_at",
     "updated_at"
    ],
    "title": "AgentOut"
   },
   "AgentUpdate": {
    "properties": {
     "name": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 128,
        "minLength": 1
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     },
     "archived": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Archived"
     }
    },
    "type": "object",
    "title": "AgentUpdate"
   },
   "AgentVersionOut": {
    "properties": {
     "version": {
      "type": "integer",
      "title": "Version"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "created_by": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Created By"
     },
     "created_by_email": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Created By Email"
     },
     "definition": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Definition"
     }
    },
    "type": "object",
    "required": [
     "version",
     "created_at",
     "created_by"
    ],
    "title": "AgentVersionOut"
   },
...
   "DefinitionIn": {
    "properties": {
     "definition": {
      "additionalProperties": true,
      "type": "object",
      "title": "Definition"
     }
    },
    "type": "object",
    "required": [
     "definition"
    ],
    "title": "DefinitionIn"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "PromptIn": {
    "properties": {
     "definition": {
      "additionalProperties": true,
      "type": "object",
      "title": "Definition"
     },
     "channel": {
      "type": "string",
      "enum": [
       "voice",
       "whatsapp"
      ],
      "title": "Channel",
      "default": "voice"
     }
    },
    "type": "object",
    "required": [
     "definition"
    ],
    "title": "PromptIn"
   },
   "PromptOut": {
    "properties": {
     "engine": {
      "type": "string",
      "title": "Engine"
     },
     "system": {
      "type": "string",
      "title": "System",
      "description": "Prompt de sistema"
     },
     "workflow": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Workflow",
      "description": "structured: la definicion que va en cada turno"
     }
    },
    "type": "object",
    "required": [
     "engine",
     "system"
    ],
    "title": "PromptOut"
   },
...
   "TemplateOut": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "engine": {
      "type": "string",
      "title": "Engine"
     },
     "agent": {
      "type": "string",
      "title": "Agent"
     },
     "voice": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Voice"
     },
     "objective": {
      "type": "string",
      "title": "Objective"
     }
    },
    "type": "object",
    "required": [
     "id",
     "engine",
     "agent",
     "voice",
     "objective"
    ],
    "title": "TemplateOut"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
   "ValidationOut": {
    "properties": {
     "valid": {
      "type": "boolean",
      "title": "Valid"
     },
     "errors": {
      "items": {
       "additionalProperties": true,
       "type": "object"
      },
      "type": "array",
      "title": "Errors",
      "default": []
     }
    },
    "type": "object",
    "required": [
     "valid"
    ],
    "title": "ValidationOut"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "description": "Token de la sesion de demo o de la llamada",
    "scheme": "bearer"
   }
  }
 }
}